+++ client/examples/examples/advanced.hpp
@@ -3,7 +3,7 @@
 // Created Date: 19/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 03/06/2021
+// Last Modified: 16/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -15,12 +15,19 @@
 
 class BurstModulation final : public autd::core::Modulation {
  public:
-  static autd::ModulationPtr create() { return std::make_shared<BurstModulation>(); }
-  autd::Error build(const autd::Configuration config) override {
-    this->buffer().resize(config.mod_buf_size(), 0);
-    this->buffer().at(0) = 0xFF;
+  static autd::ModulationPtr create(size_t buf_size = 4000, uint16_t mod_freq_div = 10) {
+    return std::make_shared<BurstModulation>(buf_size, mod_freq_div);
+  }
+  autd::Error calc() override {
+    this->buffer().resize(_buf_size, 0);
+    this->buffer().at(_buf_size - 1) = 0xFF;
     return autd::Ok(true);
   }
+
+  BurstModulation(const size_t buf_size, const uint16_t mod_freq_div) : Modulation(mod_freq_div), _buf_size(buf_size) {}
+
+ private:
+  size_t _buf_size;
 };
 
 class UniformGain final : public autd::core::Gain {
@@ -28,7 +35,7 @@ class UniformGain final : public autd::core::Gain {
   static autd::GainPtr create() { return std::make_shared<UniformGain>(); }
   autd::Error calc(const autd::GeometryPtr& geometry) override {
     for (size_t i = 0; i < geometry->num_transducers(); i++)
-      this->_data[geometry->device_idx_for_trans_idx(i)].at(i % autd::NUM_TRANS_IN_UNIT) = 0xFF00;
+      this->_data[geometry->device_idx_for_trans_idx(i)].at(i % autd::NUM_TRANS_IN_UNIT) = 0xFF80;
     this->_built = true;
     return autd::Ok(true);
   }
@@ -37,8 +44,8 @@ class UniformGain final : public autd::core::Gain {
 inline void advanced_test(autd::ControllerPtr& autd) {
   autd->silent_mode() = false;
 
-  std::vector<autd::DataArray> delays;
-  autd::DataArray delay{};
+  std::vector<std::array<uint8_t, autd::NUM_TRANS_IN_UNIT>> delays;
+  std::array<uint8_t, autd::NUM_TRANS_IN_UNIT> delay{};
   delay.fill(0);
   delay[0] = 4;  // 4 cycle = 100 us delay in 0-th transducer
   delays.emplace_back(delay);