+++ client/include/core/hardware_defined.hpp
@@ -3,7 +3,7 @@
 // Created Date: 14/04/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 22/05/2021
+// Last Modified: 16/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -30,12 +30,11 @@ constexpr auto IsMissingTransducer(T x, T y) {
 constexpr size_t FPGA_CLOCK = 20400000;
 constexpr size_t ULTRASOUND_FREQUENCY = 40000;
 
-constexpr uint16_t MOD_BUF_SIZE_MAX = 32000;
+constexpr size_t MOD_BUF_SIZE_MAX = 65536;
 constexpr size_t MOD_SAMPLING_FREQ_BASE = 40000;
 constexpr size_t MOD_FRAME_SIZE = 124;
 
-constexpr size_t POINT_SEQ_BUFFER_SIZE_MAX = 40000;
-constexpr size_t POINT_SEQ_CLK_IDX_MAX = 40000;
+constexpr size_t POINT_SEQ_BUFFER_SIZE_MAX = 65536;
 constexpr size_t POINT_SEQ_BASE_FREQ = 40000;
 
 using DataArray = std::array<uint16_t, NUM_TRANS_IN_UNIT>;
@@ -58,7 +57,6 @@ enum class COMMAND : uint8_t {
   READ_FPGA_VER_LSB = 0x04,
   READ_FPGA_VER_MSB = 0x05,
   SEQ_MODE = 0x06,
-  INIT_MOD_CLOCK = 0x07,
   CLEAR = 0x09,
   SET_DELAY = 0x0A
 };
@@ -81,23 +79,14 @@ struct SeqFocus {
   SeqFocus() = default;
 
   void set(const int32_t x, const int32_t y, const int32_t z, const uint8_t duty) {
-    set(0, x);
-    set(3, y);
-    set(6, z);
-    _buf[9] = duty;
+    _buf[0] = x & 0xFFFF;
+    _buf[1] = ((y << 2) & 0xFFFC) | ((x >> 30) & 0x0002) | ((x >> 16) & 0x0001);
+    _buf[2] = ((z << 4) & 0xFFF0) | ((y >> 28) & 0x0008) | ((y >> 14) & 0x0007);
+    _buf[3] = ((duty << 6) & 0x3FC0) | ((z >> 26) & 0x0020) | ((z >> 12) & 0x001F);
   }
 
  private:
-  uint8_t _buf[10];
-  void set(const int32_t offset, const int32_t v) {
-    uint8_t tmp = v & 0xFF;
-    _buf[offset] = tmp;
-    tmp = v >> 8 & 0xFF;
-    _buf[offset + 1] = tmp;
-    tmp = v >> 16 & 0x7F;
-    tmp |= v >> 24 & 0x80;
-    _buf[offset + 2] = tmp;
-  }
+  uint16_t _buf[4];
 };
 
 }  // namespace autd::core