+++ client/include/core/logic.hpp
@@ -3,7 +3,7 @@
 // Created Date: 11/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 02/06/2021
+// Last Modified: 16/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -78,12 +78,18 @@ class Logic {
     pack_header(COMMAND::OP, ctrl_flag, data, msg_id);
     if (mod == nullptr) return;
     auto* header = reinterpret_cast<RxGlobalHeader*>(data);
-    const auto mod_size = static_cast<uint8_t>(std::clamp(mod->buffer().size() - mod->sent(), size_t{0}, MOD_FRAME_SIZE));
-    header->mod_size = mod_size;
-    if (mod->sent() == 0) header->control_flags |= MOD_BEGIN;
+    size_t offset = 0;
+    if (mod->sent() == 0) {
+      header->control_flags |= MOD_BEGIN;
+      header->mod[0] = static_cast<uint8_t>(mod->sampling_frequency_division() & 0xFF);
+      header->mod[1] = static_cast<uint8_t>(mod->sampling_frequency_division() >> 8 & 0xFF);
+      offset += 2;
+    }
+    const auto mod_size = static_cast<uint8_t>(std::clamp(mod->buffer().size() - mod->sent(), size_t{0}, MOD_FRAME_SIZE - offset));
     if (mod->sent() + mod_size >= mod->buffer().size()) header->control_flags |= MOD_END;
+    header->mod_size = mod_size;
 
-    std::memcpy(header->mod, &mod->buffer()[mod->sent()], mod_size);
+    std::memcpy(&header->mod[offset], &mod->buffer()[mod->sent()], mod_size);
     mod->sent() += mod_size;
   }
 
@@ -120,24 +126,30 @@ class Logic {
     *size = sizeof(RxGlobalHeader) + sizeof(uint16_t) * NUM_TRANS_IN_UNIT * num_devices;
     if (seq == nullptr) return;
 
-    const auto send_size = static_cast<uint16_t>(std::clamp(seq->control_points().size() - seq->sent(), size_t{0}, size_t{49}));
-
+    auto* cursor = reinterpret_cast<uint16_t*>(data + sizeof(RxGlobalHeader));
+    size_t offset = 1;
     auto* header = reinterpret_cast<RxGlobalHeader*>(data);
-    if (seq->sent() == 0) header->control_flags |= SEQ_BEGIN;
+    if (seq->sent() == 0) {
+      header->control_flags |= SEQ_BEGIN;
+      for (size_t device = 0; device < num_devices; device++) {
+        cursor[device * NUM_TRANS_IN_UNIT + 1] = seq->sampling_frequency_division();
+        cursor[device * NUM_TRANS_IN_UNIT + 2] = static_cast<uint16_t>(geometry->wavelength() * 1000);
+      }
+      offset += 4;
+    }
+    const auto send_size = static_cast<uint16_t>(
+        std::clamp(seq->control_points().size() - seq->sent(), size_t{0}, sizeof(uint16_t) * (NUM_TRANS_IN_UNIT - offset) / sizeof(SeqFocus)));
     if (seq->sent() + send_size >= seq->control_points().size()) header->control_flags |= SEQ_END;
 
-    auto* cursor = reinterpret_cast<uint16_t*>(data + sizeof(RxGlobalHeader));
-    const auto fixed_num_unit = 255 / geometry->wavelength();
+    const auto fixed_num_unit = 255.0 / geometry->wavelength();
     for (size_t device = 0; device < num_devices; device++) {
       cursor[0] = send_size;
-      cursor[1] = seq->sampling_frequency_division();
-      cursor[2] = static_cast<uint16_t>(geometry->wavelength() * 1000);
-      auto* focus_cursor = reinterpret_cast<SeqFocus*>(&cursor[4]);
+      auto* focus_cursor = reinterpret_cast<SeqFocus*>(&cursor[offset]);
       for (size_t i = 0; i < send_size; i++) {
         auto v64 = geometry->local_position(device, seq->control_points()[seq->sent() + i]);
-        const auto x = static_cast<uint32_t>(static_cast<int32_t>(v64.x() * fixed_num_unit));
-        const auto y = static_cast<uint32_t>(static_cast<int32_t>(v64.y() * fixed_num_unit));
-        const auto z = static_cast<uint32_t>(static_cast<int32_t>(v64.z() * fixed_num_unit));
+        const auto x = static_cast<int32_t>(v64.x() * fixed_num_unit);
+        const auto y = static_cast<int32_t>(v64.y() * fixed_num_unit);
+        const auto z = static_cast<int32_t>(v64.z() * fixed_num_unit);
         focus_cursor->set(x, y, z, 0xFF);
         focus_cursor++;
       }
@@ -145,36 +157,17 @@ class Logic {
     }
     seq->sent() += send_size;
   }
-
-  /**
-   * \brief Pack data body to synchronize devices
-   * \param config Configuration for Modulation
-   * \param num_devices number of devices
-   * \param[out] data pointer to transmission data
-   * \param[out] size size to send
-   */
-  static void pack_sync_body(const Configuration config, const size_t num_devices, uint8_t* data, size_t* const size) {
-    *size = sizeof(RxGlobalHeader) + sizeof(uint16_t) * NUM_TRANS_IN_UNIT * num_devices;
-
-    auto* cursor = reinterpret_cast<uint16_t*>(data + sizeof(RxGlobalHeader));
-    for (size_t i = 0; i < num_devices; i++) {
-      cursor[0] = config.mod_buf_size();
-      cursor[1] = config.mod_sampling_freq_div();
-      cursor += NUM_TRANS_IN_UNIT;
-    }
-  }
-
   /**
    * \brief Pack data body to set output delay
    * \param delay delay data of each transducer
    * \param[out] data pointer to transmission data
    * \param[out] size size to send
    */
-  static void pack_delay_body(const std::vector<DataArray>& delay, uint8_t* data, size_t* const size) {
+  static void pack_delay_body(const std::vector<std::array<uint8_t, NUM_TRANS_IN_UNIT>>& delay, uint8_t* data, size_t* const size) {
     *size = sizeof(RxGlobalHeader) + sizeof(uint16_t) * NUM_TRANS_IN_UNIT * delay.size();
     auto* cursor = reinterpret_cast<uint16_t*>(data + sizeof(RxGlobalHeader));
     for (auto&& d : delay) {
-      std::memcpy(cursor, &d[0], NUM_TRANS_IN_UNIT);
+      for (size_t i = 0; i < NUM_TRANS_IN_UNIT; i++) cursor[i] = d[i];
       cursor += NUM_TRANS_IN_UNIT;
     }
   }