+++ client/include/core/modulation.hpp
@@ -3,7 +3,7 @@
 // Created Date: 11/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 08/06/2021
+// Last Modified: 16/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -14,7 +14,7 @@
 #include <memory>
 #include <vector>
 
-#include "configuration.hpp"
+#include "hardware_defined.hpp"
 #include "result.hpp"
 #include "utils.hpp"
 
@@ -28,7 +28,8 @@ using ModulationPtr = std::shared_ptr<Modulation>;
  */
 class Modulation {
  public:
-  Modulation() noexcept : _sent(0) {}
+  Modulation() noexcept : Modulation(10) {}
+  explicit Modulation(const uint16_t freq_div) noexcept : _freq_div(freq_div), _sent(0) {}
   virtual ~Modulation() = default;
   Modulation(const Modulation& v) noexcept = default;
   Modulation& operator=(const Modulation& obj) = default;
@@ -48,7 +49,7 @@ class Modulation {
    */
   static ModulationPtr create(const uint8_t duty = 0xFF) {
     auto mod = std::make_shared<Modulation>();
-    mod->_buffer.resize(MOD_FRAME_SIZE, duty);
+    mod->_buffer.resize(1, duty);
     return mod;
   }
 
@@ -59,13 +60,23 @@ class Modulation {
   static ModulationPtr create(const double amp) { return create(to_duty(amp)); }
 
   /**
-   * \brief Build modulation data with Configuration
-   * \param config Configuration
+   * \brief Calculate modulation data
    * \return ok(whether succeeded), or err(error message) if unrecoverable error is occurred
    */
-  [[nodiscard]] virtual Error build(Configuration config) {
-    (void)config;
-    return Ok(true);
+  [[nodiscard]] virtual Error calc() { return Ok(true); }
+
+  /**
+   * \brief Build modulation data
+   * \return ok(whether succeeded), or err(error message) if unrecoverable error is occurred
+   */
+  [[nodiscard]] Error build() {
+    bool success;
+    if (auto res = this->calc(); res.is_err())
+      return res;
+    else
+      success = res.unwrap();
+    if (this->_buffer.size() > MOD_BUF_SIZE_MAX) return Err(std::string("Modulation buffer overflow"));
+    return Ok(success);
   }
 
   /**
@@ -78,7 +89,18 @@ class Modulation {
    */
   std::vector<uint8_t>& buffer() { return _buffer; }
 
+  /**
+   * \brief modulation sampling frequency division
+   */
+  [[nodiscard]] uint16_t sampling_frequency_division() const noexcept { return _freq_div; }
+
+  /**
+   * \brief modulation sampling frequency
+   */
+  [[nodiscard]] double sampling_freq() const noexcept { return static_cast<double>(MOD_SAMPLING_FREQ_BASE) / static_cast<double>(this->_freq_div); }
+
  protected:
+  uint16_t _freq_div;
   size_t _sent;
   std::vector<uint8_t> _buffer;
 };